# src/doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000/api"
info:
  title: Portale Studenti Sapienza (Vulnerabile per Test)
  description: API RESTful per la gestione della carriera universitaria (Profilo, Libretto, Prenotazioni), con vulnerabilità SQL Injection intenzionali per scopi didattici.
  version: 1.0.0
paths:
  /login:
    post:
      tags: ["Autenticazione"]
      operationId: doLogin
      summary: Effettua il login al portale (VULNERABILE a SQLi)
      description: |-
        Endpoint per l'autenticazione dello studente.
        **Vulnerabilità:** Intenzionalmente suscettibile ad attacchi di Tautologia SQL Injection nel campo password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - matricola
                - password
              properties:
                matricola:
                  type: integer
                  description: La matricola dello studente
                  example: 2112808
                password:
                  type: string
                  description: La password (o payload SQLi)
                  example: "Sapienza123$"
      responses:
        "200":
          description: Login completato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  /profilo/{matricola}:
    parameters:
      - $ref: "#/components/parameters/PathMatricola"
    get:
      tags: ["Dati Studente"]
      operationId: getProfilo
      summary: Recupera i dati anagrafici dello studente
      description: Restituisce nome, cognome, indirizzo, email e data di nascita legati a una specifica matricola.
      responses:
        "200":
          description: Profilo recuperato con successo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfiloResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  /libretto/{matricola}:
    parameters:
      - $ref: "#/components/parameters/PathMatricola"
    get:
      tags: ["Dati Studente"]
      operationId: getLibretto
      summary: Recupera gli esami superati
      description: Restituisce la lista degli esami registrati a libretto con relativi voti e CFU per il calcolo della media.
      responses:
        "200":
          description: Libretto recuperato con successo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibrettoResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  /prenotazioni:
    get:
      tags: ["Gestione Appelli"]
      operationId: searchPrenotazioni
      summary: Cerca le prenotazioni dello studente (VULNERABILE a SQLi)
      description: |-
        Recupera gli appelli a cui lo studente si è prenotato, permettendo la ricerca per nome insegnamento.
        **Vulnerabilità:** Il parametro di ricerca `q` è intenzionalmente vulnerabile a Error-Based SQL Injection.
      parameters:
        - name: matricola
          in: query
          required: true
          description: Matricola dell'utente loggato per filtrare i risultati
          schema:
            type: integer
        - name: q
          in: query
          required: false
          description: Testo per filtrare gli insegnamenti
          schema:
            type: string
      responses:
        "200":
          description: Ricerca completata con successo (o errore SQL stampato a video)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrenotazioniResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }

components:
  schemas:
    # --- Modelli di Risposta ---
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
        matricola:
          type: integer
        nome:
          type: string
        cognome:
          type: string

    ProfiloResponse:
      type: object
      properties:
        studente:
          type: array
          items:
            type: object
            properties:
              matricola: { type: integer }
              nome: { type: string }
              cognome: { type: string }
              data_di_nascita: { type: string }
              indirizzo: { type: string }
              email: { type: string }
        error:
          type: string

    LibrettoResponse:
      type: object
      properties:
        libretto:
          type: array
          items:
            type: object
            properties:
              matricola: { type: integer }
              id_appello: { type: integer }
              insegnamento: { type: string }
              data_registrazione: { type: string }
              voto: { type: integer }
              cfu: { type: integer }
        error:
          type: string

    PrenotazioniResponse:
      type: object
      properties:
        prenotazione:
          type: array
          items:
            type: object
            properties:
              id_prenotazione: { type: integer }
              id_appello: { type: integer }
              insegnamento: { type: string }
              data_esame: { type: string }
              aula: { type: string }
              stato: { type: string }
        error:
          type: string
          description: Contiene l'eventuale errore SQL stampato a video per l'attacco Error-Based.

  responses:
    BadRequest:
      description: Parametri mancanti o non validi.
    InternalServerError:
      description: Errore interno del server o del database.

  parameters:
    PathMatricola:
      name: matricola
      in: path
      required: true
      description: La matricola dello studente
      schema:
        type: integer